+++ packages/adminui-vue/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.16]
+### Fixed
+ - fix #63 - placeholder translation for u-auto-field of Date/DateTime type
+ 
 ## [1.6.14]
 ### Added
  - exports `dialog`, `dialogError`, `dialogInfo`, `dialogYesNo` functions from `adminui-vue`. To be used outside Vue instance: